feat: Add configFormService for refacto display config form
