+++ alma/controllers/hook/GetContentHookController.php
@@ -33,23 +33,14 @@ use Alma\API\Lib\IntegrationsConfigurationsUtils;
 use Alma\API\RequestError;
 use Alma\PrestaShop\Builders\Helpers\ApiHelperBuilder;
 use Alma\PrestaShop\Builders\Helpers\ContextHelperBuilder;
-use Alma\PrestaShop\Builders\Helpers\CustomFieldHelperBuilder;
 use Alma\PrestaShop\Builders\Helpers\PriceHelperBuilder;
 use Alma\PrestaShop\Builders\Helpers\SettingsHelperBuilder;
 use Alma\PrestaShop\Builders\Helpers\ShareOfCheckoutHelperBuilder;
-use Alma\PrestaShop\Builders\Models\MediaHelperBuilder;
 use Alma\PrestaShop\Exceptions\MissingParameterException;
-use Alma\PrestaShop\Factories\ClientFactory;
 use Alma\PrestaShop\Forms\ApiAdminFormBuilder;
-use Alma\PrestaShop\Forms\CartEligibilityAdminFormBuilder;
-use Alma\PrestaShop\Forms\DebugAdminFormBuilder;
 use Alma\PrestaShop\Forms\ExcludedCategoryAdminFormBuilder;
 use Alma\PrestaShop\Forms\InpageAdminFormBuilder;
 use Alma\PrestaShop\Forms\PaymentButtonAdminFormBuilder;
-use Alma\PrestaShop\Forms\PaymentOnTriggeringAdminFormBuilder;
-use Alma\PrestaShop\Forms\PnxAdminFormBuilder;
-use Alma\PrestaShop\Forms\ProductEligibilityAdminFormBuilder;
-use Alma\PrestaShop\Forms\RefundAdminFormBuilder;
 use Alma\PrestaShop\Forms\ShareOfCheckoutAdminFormBuilder;
 use Alma\PrestaShop\Helpers\ApiHelper;
 use Alma\PrestaShop\Helpers\ApiKeyHelper;
@@ -57,13 +48,11 @@ use Alma\PrestaShop\Helpers\ClientHelper;
 use Alma\PrestaShop\Helpers\CmsDataHelper;
 use Alma\PrestaShop\Helpers\ConstantsHelper;
 use Alma\PrestaShop\Helpers\ContextHelper;
-use Alma\PrestaShop\Helpers\CustomFieldsHelper;
-use Alma\PrestaShop\Helpers\MediaHelper;
 use Alma\PrestaShop\Helpers\PriceHelper;
 use Alma\PrestaShop\Helpers\SettingsHelper;
 use Alma\PrestaShop\Hooks\AdminHookController;
 use Alma\PrestaShop\Logger;
-use Alma\PrestaShop\Model\ClientModel;
+use Alma\PrestaShop\Services\ConfigFormService;
 
 final class GetContentHookController extends AdminHookController
 {
@@ -87,16 +76,6 @@ final class GetContentHookController extends AdminHookController
      * @var PriceHelper
      */
     protected $priceHelper;
-
-    /**
-     * @var CustomFieldsHelper
-     */
-    protected $customFieldsHelper;
-
-    /**
-     * @var MediaHelper
-     */
-    protected $mediaHelper;
 
     /**
      * @var bool
@@ -165,9 +144,9 @@ final class GetContentHookController extends AdminHookController
         'ALMA_DESCRIPTION_TRIGGER' => 'none',
     ];
     /**
-     * @var \Alma\PrestaShop\Factories\ClientFactory
+     * @var \Alma\PrestaShop\Services\ConfigFormService
      */
-    protected $clientFactory;
+    protected $configFormService;
 
     /**
      * GetContentHook Controller construct.
@@ -187,16 +166,13 @@ final class GetContentHookController extends AdminHookController
         $priceHelperBuilder = new PriceHelperBuilder();
         $this->priceHelper = $priceHelperBuilder->getInstance();
 
-        $customFieldHelperBuilder = new CustomFieldHelperBuilder();
-        $this->customFieldsHelper = $customFieldHelperBuilder->getInstance();
-
-        $mediaHelperBuilder = new MediaHelperBuilder();
-        $this->mediaHelper = $mediaHelperBuilder->getInstance();
-
         $contextHelperBuilder = new ContextHelperBuilder();
         $this->contextHelper = $contextHelperBuilder->getInstance();
 
-        $this->clientFactory = new ClientFactory();
+        $this->configFormService = new ConfigFormService(
+            $module,
+            $this->settingsHelper
+        );
 
         $this->hasKey = false;
 
@@ -592,421 +568,6 @@ final class GetContentHookController extends AdminHookController
     }
 
     /**
-     * @return string
-     *
-     * @throws \Exception
-     */
-    public function renderForm()
-    {
-        // Try to get merchant from configured API key/mode
-        $merchant = null;
-
-        try {
-            $merchant = $this->apiHelper->getMerchant();
-        } catch (\Exception $e) {
-            Logger::instance()->error($e->getMessage());
-        }
-
-        $extraMessage = null;
-
-        $needsKeys = $this->needsAPIKey();
-
-        if ($needsKeys
-            && !\Tools::isSubmit('alma_config_form')
-        ) {
-            $this->context->smarty->clearAllAssign();
-
-            $this->assignSmartyAlertClasses();
-            $this->context->smarty->assign('tip', 'fill_api_keys');
-            $this->context->smarty->assign('suggestPSAccounts', false);
-            $this->context->smarty->assign('hasPSAccounts', false);
-
-            $extraMessage = $this->module->display($this->module->file, 'getContent.tpl');
-        }
-
-        $this->assignSmartyAlertClasses();
-
-        $feePlans = $this->getFeePlans();
-
-        $feePlansOrdered = $this->getPlansForForms($feePlans, $merchant);
-        $installmentsPlans = json_decode(SettingsHelper::getFeePlans());
-
-        $fieldsForms = $this->buildForms($needsKeys, $feePlansOrdered, $installmentsPlans);
-
-        $helper = $this->getHelperForm();
-
-        if ($merchant) {
-            $sortOrder = 1;
-            foreach ($feePlans as $feePlan) {
-                $key = $this->settingsHelper->keyForFeePlan($feePlan);
-
-                $helper->fields_value["ALMA_{$key}_ENABLED_ON"] = isset($installmentsPlans->$key->enabled)
-                    ? $installmentsPlans->$key->enabled
-                    : 0;
-
-                $minAmount = isset($installmentsPlans->$key->min)
-                    ? $installmentsPlans->$key->min
-                    : $feePlan->min_purchase_amount;
-
-                $helper->fields_value["ALMA_{$key}_MIN_AMOUNT"] = (int) round(
-                    $this->priceHelper->convertPriceFromCents($minAmount)
-                );
-                $maxAmount = isset($installmentsPlans->$key->max)
-                    ? $installmentsPlans->$key->max
-                    : $feePlan->max_purchase_amount;
-
-                $helper->fields_value["ALMA_{$key}_MAX_AMOUNT"] = (int) $this->priceHelper->convertPriceFromCents($maxAmount);
-
-                $order = isset($installmentsPlans->$key->order)
-                    ? $installmentsPlans->$key->order
-                    : $sortOrder;
-
-                $helper->fields_value["ALMA_{$key}_SORT_ORDER"] = $order;
-
-                ++$sortOrder;
-            }
-        }
-
-        return $extraMessage . $helper->generateForm($fieldsForms);
-    }
-
-    /**
-     * @param bool $needsKeys
-     * @param array $feePlansOrdered
-     * @param array $installmentsPlans
-     *
-     * @return array
-     */
-    protected function buildForms($needsKeys, $feePlansOrdered, $installmentsPlans)
-    {
-        $iconPath = $this->mediaHelper->getIconPathAlmaTiny();
-        $fieldsForms = [];
-
-        if (!$needsKeys) {
-            $pnxBuilder = new PnxAdminFormBuilder(
-                $this->module,
-                $this->context,
-                $iconPath,
-                ['feePlans' => $feePlansOrdered, 'installmentsPlans' => $installmentsPlans]
-            );
-            if ($pnxBuilder) {
-                $fieldsForms[] = $pnxBuilder->build();
-            }
-
-            $productBuilder = new ProductEligibilityAdminFormBuilder($this->module, $this->context, $iconPath);
-            $fieldsForms[] = $productBuilder->build();
-
-            $cartBuilder = new CartEligibilityAdminFormBuilder($this->module, $this->context, $iconPath);
-            $fieldsForms[] = $cartBuilder->build();
-
-            $paymentBuilder = new PaymentButtonAdminFormBuilder($this->module, $this->context, $iconPath);
-            $fieldsForms[] = $paymentBuilder->build();
-
-            $excludedBuilder = new ExcludedCategoryAdminFormBuilder($this->module, $this->context, $iconPath);
-            $fieldsForms[] = $excludedBuilder->build();
-
-            $refundBuilder = new RefundAdminFormBuilder($this->module, $this->context, $iconPath);
-            $fieldsForms[] = $refundBuilder->build();
-
-            if (!SettingsHelper::shouldHideShareOfCheckoutForm()) {
-                $shareOfCheckoutBuilder = new ShareOfCheckoutAdminFormBuilder($this->module, $this->context, $iconPath);
-                $fieldsForms[] = $shareOfCheckoutBuilder->build();
-            }
-
-            $inpageBuilder = new InpageAdminFormBuilder($this->module, $this->context, $iconPath);
-            $fieldsForms[] = $inpageBuilder->build();
-        }
-
-        if ($this->settingsHelper->isPaymentTriggerEnabledByState()) {
-            $triggerBuilder = new PaymentOnTriggeringAdminFormBuilder($this->module, $this->context, $iconPath);
-            $fieldsForms[] = $triggerBuilder->build();
-        }
-
-        $apiBuilder = new ApiAdminFormBuilder($this->module, $this->context, $iconPath, ['needsAPIKey' => $needsKeys]);
-        $fieldsForms[] = $apiBuilder->build();
-
-        $debugBuilder = new DebugAdminFormBuilder($this->module, $this->context, $iconPath);
-        $fieldsForms[] = $debugBuilder->build();
-
-        return $fieldsForms;
-    }
-
-    /**
-     * @param $feePlansOrdered
-     *
-     * @return array|array[]
-     */
-    protected function getPnxAdminFormBuilder($feePlansOrdered)
-    {
-        $installmentsPlans = json_decode(SettingsHelper::getFeePlans());
-
-        $pnxBuilder = new PnxAdminFormBuilder(
-            $this->module,
-            $this->context,
-            $this->mediaHelper->getIconPathAlmaTiny(),
-            ['feePlans' => $feePlansOrdered, 'installmentsPlans' => $installmentsPlans]
-        );
-
-        return $pnxBuilder->build();
-    }
-
-    /**
-     * @return array
-     */
-    protected function getProductEligibilityAdminForm()
-    {
-        $productBuilder = new ProductEligibilityAdminFormBuilder($this->module, $this->context, $this->mediaHelper->getIconPathAlmaTiny());
-
-        return $productBuilder->build();
-    }
-
-    /**
-     * @return array|array[]
-     */
-    protected function getCartEligibilityAdminForm()
-    {
-        $cartBuilder = new CartEligibilityAdminFormBuilder($this->module, $this->context, $this->mediaHelper->getIconPathAlmaTiny());
-
-        return $cartBuilder->build();
-    }
-
-    /**
-     * @return array|array[]
-     */
-    protected function getPaymentButtonAdminForm()
-    {
-        $paymentBuilder = new PaymentButtonAdminFormBuilder($this->module, $this->context, $this->mediaHelper->getIconPathAlmaTiny());
-
-        return $paymentBuilder->build();
-    }
-
-    /**
-     * @return array|array[]
-     */
-    protected function getExcludedCategoryAdminForm()
-    {
-        $excludedBuilder = new ExcludedCategoryAdminFormBuilder($this->module, $this->context, $this->mediaHelper->getIconPathAlmaTiny());
-
-        return $excludedBuilder->build();
-    }
-
-    /**
-     * @return array|array[]
-     */
-    protected function getRefundAdminForm()
-    {
-        $refundBuilder = new RefundAdminFormBuilder($this->module, $this->context, $this->mediaHelper->getIconPathAlmaTiny());
-
-        return $refundBuilder->build();
-    }
-
-    /**
-     * @return array|array[]
-     */
-    protected function getShareOfCheckoutAdminForm()
-    {
-        $shareOfCheckoutBuilder = new ShareOfCheckoutAdminFormBuilder($this->module, $this->context, $this->mediaHelper->getIconPathAlmaTiny());
-
-        return $shareOfCheckoutBuilder->build();
-    }
-
-    /**
-     * @return array|array[]
-     */
-    protected function getInpageAdminForm()
-    {
-        $inpageBuilder = new InpageAdminFormBuilder($this->module, $this->context, $this->mediaHelper->getIconPathAlmaTiny());
-
-        return $inpageBuilder->build();
-    }
-
-    /**
-     * @return array|array[]
-     */
-    protected function getPaymentOnTriggeringAdminForm()
-    {
-        $triggerBuilder = new PaymentOnTriggeringAdminFormBuilder($this->module, $this->context, $this->mediaHelper->getIconPathAlmaTiny());
-
-        return $triggerBuilder->build();
-    }
-
-    /**
-     * @param bool $needsKeys
-     *
-     * @return array|array[]
-     */
-    protected function getApiAdminFormBuilder($needsKeys = false)
-    {
-        $apiBuilder = new ApiAdminFormBuilder($this->module, $this->context, $this->mediaHelper->getIconPathAlmaTiny(), ['needsAPIKey' => $needsKeys]);
-
-        return $apiBuilder->build();
-    }
-
-    /**
-     * @return array|array[]
-     */
-    protected function getDebugAdminFormBuilder()
-    {
-        $debugBuilder = new DebugAdminFormBuilder($this->module, $this->context, $this->mediaHelper->getIconPathAlmaTiny());
-
-        return $debugBuilder->build();
-    }
-
-    /**
-     * @return array
-     */
-    protected function getFieldsValueForForm()
-    {
-        return [
-            'ALMA_LIVE_API_KEY' => SettingsHelper::getLiveKey(),
-            'ALMA_TEST_API_KEY' => SettingsHelper::getTestKey(),
-            'ALMA_API_MODE' => SettingsHelper::getActiveMode(),
-            PaymentButtonAdminFormBuilder::ALMA_PAY_NOW_BUTTON_TITLE => $this->customFieldsHelper->getValue(
-                PaymentButtonAdminFormBuilder::ALMA_PAY_NOW_BUTTON_TITLE
-            ),
-            PaymentButtonAdminFormBuilder::ALMA_PAY_NOW_BUTTON_DESC => $this->customFieldsHelper->getValue(
-                PaymentButtonAdminFormBuilder::ALMA_PAY_NOW_BUTTON_DESC
-            ),
-            PaymentButtonAdminFormBuilder::ALMA_PNX_BUTTON_TITLE => $this->customFieldsHelper->getValue(
-                PaymentButtonAdminFormBuilder::ALMA_PNX_BUTTON_TITLE
-            ),
-            PaymentButtonAdminFormBuilder::ALMA_PNX_BUTTON_DESC => $this->customFieldsHelper->getValue(
-                PaymentButtonAdminFormBuilder::ALMA_PNX_BUTTON_DESC
-            ),
-            PaymentButtonAdminFormBuilder::ALMA_DEFERRED_BUTTON_TITLE => $this->customFieldsHelper->getValue(
-                PaymentButtonAdminFormBuilder::ALMA_DEFERRED_BUTTON_TITLE
-            ),
-            PaymentButtonAdminFormBuilder::ALMA_DEFERRED_BUTTON_DESC => $this->customFieldsHelper->getValue(
-                PaymentButtonAdminFormBuilder::ALMA_DEFERRED_BUTTON_DESC
-            ),
-            PaymentButtonAdminFormBuilder::ALMA_PNX_AIR_BUTTON_TITLE => $this->customFieldsHelper->getValue(
-                PaymentButtonAdminFormBuilder::ALMA_PNX_AIR_BUTTON_TITLE
-            ),
-            PaymentButtonAdminFormBuilder::ALMA_PNX_AIR_BUTTON_DESC => $this->customFieldsHelper->getValue(
-                PaymentButtonAdminFormBuilder::ALMA_PNX_AIR_BUTTON_DESC
-            ),
-            InpageAdminFormBuilder::ALMA_ACTIVATE_INPAGE . '_ON' => $this->settingsHelper->isInPageEnabled(),
-            InpageAdminFormBuilder::ALMA_INPAGE_PAYMENT_BUTTON_SELECTOR => $this->settingsHelper->getKey(InpageAdminFormBuilder::ALMA_INPAGE_PAYMENT_BUTTON_SELECTOR, InpageAdminFormBuilder::ALMA_INPAGE_DEFAULT_VALUE_PAYMENT_BUTTON_SELECTOR),
-            InpageAdminFormBuilder::ALMA_INPAGE_PLACE_ORDER_BUTTON_SELECTOR => $this->settingsHelper->getKey(InpageAdminFormBuilder::ALMA_INPAGE_PLACE_ORDER_BUTTON_SELECTOR, InpageAdminFormBuilder::ALMA_INPAGE_DEFAULT_VALUE_PLACE_ORDER_BUTTON_SELECTOR),
-            'ALMA_SHOW_DISABLED_BUTTON' => SettingsHelper::showDisabledButton(),
-            'ALMA_SHOW_ELIGIBILITY_MESSAGE_ON' => SettingsHelper::showEligibilityMessage(),
-            'ALMA_CART_WDGT_NOT_ELGBL_ON' => SettingsHelper::showCartWidgetIfNotEligible(),
-            'ALMA_PRODUCT_WDGT_NOT_ELGBL_ON' => SettingsHelper::showProductWidgetIfNotEligible(),
-            'ALMA_CATEGORIES_WDGT_NOT_ELGBL_ON' => SettingsHelper::showCategoriesWidgetIfNotEligible(),
-            'ALMA_ACTIVATE_LOGGING_ON' => (bool) SettingsHelper::canLog(),
-            'ALMA_SHARE_OF_CHECKOUT_STATE_ON' => SettingsHelper::getShareOfCheckoutStatus(),
-            'ALMA_SHARE_OF_CHECKOUT_DATE' => SettingsHelper::getCurrentTimestamp(),
-            'ALMA_STATE_REFUND' => SettingsHelper::getRefundState(),
-            'ALMA_STATE_REFUND_ENABLED_ON' => SettingsHelper::isRefundEnabledByState(),
-            'ALMA_STATE_TRIGGER' => SettingsHelper::getPaymentTriggerState(),
-            'ALMA_PAYMENT_ON_TRIGGERING_ENABLED_ON' => $this->settingsHelper->isPaymentTriggerEnabledByState(),
-            'ALMA_DESCRIPTION_TRIGGER' => SettingsHelper::getKeyDescriptionPaymentTrigger(),
-            'ALMA_NOT_ELIGIBLE_CATEGORIES' => $this->customFieldsHelper->getValue(
-                ExcludedCategoryAdminFormBuilder::ALMA_NOT_ELIGIBLE_CATEGORIES
-            ),
-            'ALMA_SHOW_PRODUCT_ELIGIBILITY_ON' => SettingsHelper::showProductEligibility(),
-            'ALMA_PRODUCT_PRICE_SELECTOR' => SettingsHelper::getProductPriceQuerySelector(),
-            'ALMA_WIDGET_POSITION_SELECTOR' => SettingsHelper::getProductWidgetPositionQuerySelector(),
-            'ALMA_WIDGET_POSITION_CUSTOM' => SettingsHelper::isWidgetCustomPosition(),
-            'ALMA_CART_WDGT_POS_SELECTOR' => SettingsHelper::getCartWidgetPositionQuerySelector(),
-            'ALMA_CART_WIDGET_POSITION_CUSTOM' => SettingsHelper::isCartWidgetCustomPosition(),
-            'ALMA_PRODUCT_ATTR_SELECTOR' => SettingsHelper::getProductAttrQuerySelector(),
-            'ALMA_PRODUCT_ATTR_RADIO_SELECTOR' => SettingsHelper::getProductAttrRadioQuerySelector(),
-            'ALMA_PRODUCT_COLOR_PICK_SELECTOR' => SettingsHelper::getProductColorPickQuerySelector(),
-            'ALMA_PRODUCT_QUANTITY_SELECTOR' => SettingsHelper::getProductQuantityQuerySelector(),
-            '_api_only' => true,
-        ];
-    }
-
-    /**
-     * @param array|null $feePlans
-     * @param Merchant|null $merchant
-     *
-     * @return array
-     */
-    protected function getPlansForForms($feePlans, $merchant = null)
-    {
-        $feePlansOrdered = [];
-
-        if ($merchant) {
-            // sort fee plans by pnx then by pay later duration
-            $feePlanDeferred = [];
-
-            foreach ($feePlans as $feePlan) {
-                if (!$this->settingsHelper->isDeferred($feePlan)) {
-                    $feePlansOrdered[$feePlan->installments_count] = $feePlan;
-                    continue;
-                }
-
-                $duration = $this->settingsHelper->getDuration($feePlan);
-                $feePlanDeferred[$feePlan->installments_count . $duration] = $feePlan;
-            }
-
-            ksort($feePlanDeferred);
-            $feePlansOrdered = array_merge($feePlansOrdered, $feePlanDeferred);
-        }
-
-        return $feePlansOrdered;
-    }
-
-    private function assignSmartyAlertClasses($level = 'danger')
-    {
-        $token = \Tools::getAdminTokenLite('AdminModules');
-        $href = $this->context->link->getAdminLink('AdminParentModulesSf', $token);
-
-        if (version_compare(_PS_VERSION_, '1.6', '<')) {
-            $this->context->smarty->assign([
-                'validation_error_classes' => 'alert',
-                'tip_classes' => 'conf',
-                'success_classes' => 'conf',
-            ]);
-        } else {
-            $this->context->smarty->assign([
-                'validation_error_classes' => "alert alert-$level",
-                'tip_classes' => 'alert alert-info',
-                'success_classes' => 'alert alert-success',
-                'breadcrumbs2' => [
-                    'container' => [
-                        'name' => $this->module->l('Modules', 'GetContentHookController'),
-                        'href' => $href,
-                    ],
-                    'tab' => [
-                        'name' => $this->module->l('Module Manager', 'GetContentHookController'),
-                        'href' => $href,
-                    ],
-                ],
-                'quick_access_current_link_name' => $this->module->l('Module Manager - List', 'GetContentHookController'),
-                'quick_access_current_link_icon' => 'icon-AdminParentModulesSf',
-                'token' => $token,
-                'host_mode' => 0,
-            ]);
-        }
-    }
-
-    /**
-     * @return bool
-     *
-     * @throws \Exception
-     */
-    public function needsAPIKey()
-    {
-        $key = trim(SettingsHelper::getActiveAPIKey());
-
-        return '' == $key || null == $key;
-    }
-
-    private function renderTemplate($formFields, $feePlans = [])
-    {
-        $helperForm = $this->getHelperForm();
-
-        if ($feePlans) {
-            $helperForm->fields_value = array_merge($helperForm->fields_value, $this->getFieldsValueFromFeePlans($feePlans));
-        }
-
-        return $helperForm->generateForm($formFields);
-    }
-
-    /**
      * @param $params
      *
      * @return string
@@ -1017,91 +578,16 @@ final class GetContentHookController extends AdminHookController
     {
         $assignSmartyKeys = $this->assignSmartyKeys();
         $messages = [];
-        $feePlans = [];
-
-        /** @var \Alma\API\Client|null $almaClient */
-        $almaClient = $this->clientFactory->get();
 
         if (\Tools::isSubmit('alma_config_form')) {
             $messages = $this->processConfiguration();
         }
 
-        if ($almaClient) {
-            $clientModel = new ClientModel($almaClient);
-            $merchant = $clientModel->getMerchantMe();
-            $feePlans = $clientModel->getMerchantFeePlans();
-
-            $formFields = [];
-
-            if (!$this->needsAPIKey()) {
-                $feePlansOrdered = $this->getPlansForForms($feePlans, $merchant);
-                $formFields[] = $this->getPnxAdminFormBuilder($feePlansOrdered);
-                $formFields[] = $this->getProductEligibilityAdminForm();
-                $formFields[] = $this->getCartEligibilityAdminForm();
-                $formFields[] = $this->getPaymentButtonAdminForm();
-                $formFields[] = $this->getExcludedCategoryAdminForm();
-                $formFields[] = $this->getRefundAdminForm();
-                if (!SettingsHelper::shouldHideShareOfCheckoutForm()) {
-                    $formFields[] = $this->getShareOfCheckoutAdminForm();
-                }
-                $formFields[] = $this->getInpageAdminForm();
-                if ($this->settingsHelper->isPaymentTriggerEnabledByState()) {
-                    $formFields[] = $this->getPaymentOnTriggeringAdminForm();
-                }
-            }
-        }
-        $formFields[] = $this->getApiAdminFormBuilder(true);
-        $formFields[] = $this->getDebugAdminFormBuilder();
-
-        $assignSmartyKeys['form'] = $this->renderTemplate($formFields, $feePlans);
+        $assignSmartyKeys['form'] = $this->configFormService->getRenderHtml();
         $assignSmartyKeys['error_messages'] = $messages;
         $this->context->smarty->assign($assignSmartyKeys);
 
         return $this->module->display($this->module->file, 'configurationContent.tpl');
-
-        //Old code
-//        $this->context->smarty->assign([
-//            'hasPSAccounts' => $params['hasPSAccounts'],
-//            'updated' => true,
-//            'suggestPSAccounts' => $params['suggestPSAccounts'],
-//        ]);
-//
-//        $this->assignSmartyAlertClasses();
-//
-//        $messages = null;
-//
-//        if (\Tools::isSubmit('alma_config_form')) {
-//            $messages = $this->processConfiguration();
-//        } elseif (!$this->needsAPIKey()) {
-//            $messages = $this->credentialsError(
-//                SettingsHelper::getLiveKey(),
-//                SettingsHelper::getTestKey()
-//            );
-//
-//            if ($messages) {
-//                $messages = $messages['message'];
-//            } else {
-//                $this->hasKey = true;
-//            }
-//        }
-//        $this->context->smarty->assign([
-//            'hasKey' => $this->hasKey,
-//        ]);
-//
-//        if (!$messages) {
-//            $this->context->smarty->assign([
-//                'updated' => false,
-//            ]);
-//            $messages = '';
-//        }
-//
-//        if ($params['hasPSAccounts'] || $params['suggestPSAccounts']) {
-//            $messages = $this->module->display($this->module->file, 'getContent.tpl');
-//        }
-//
-//        $htmlForm = $this->renderForm();
-//
-//        return $messages . $htmlForm;
     }
 
     /**
@@ -1219,75 +705,5 @@ final class GetContentHookController extends AdminHookController
 
             $this->updateSettingsValue($key, $value);
         }
-    }
-
-    /**
-     * @return \HelperForm
-     */
-    public function getHelperForm()
-    {
-        $helper = new \HelperForm();
-        $helper->module = $this->module;
-        $helper->table = 'alma_config';
-        $helper->default_form_language = (int) \Configuration::get('PS_LANG_DEFAULT');
-        $helper->allow_employee_form_lang = (int) \Configuration::get('PS_BO_ALLOW_EMPLOYEE_FORM_LANG');
-        $helper->submit_action = 'alma_config_form';
-
-        if (version_compare(_PS_VERSION_, '1.6', '<')) {
-            $helper->base_folder = 'helpers/form/15/';
-            $this->context->controller->addCss(_MODULE_DIR_ . $this->module->name . '/views/css/admin/tabs.css');
-        }
-        $helper->currentIndex = $this->context->link->getAdminLink('AdminModules', false)
-            . '&configure=' . $this->module->name
-            . '&tab_module=' . $this->module->tab
-            . '&module_name=' . $this->module->name;
-
-        $helper->token = \Tools::getAdminTokenLite('AdminModules');
-        $helper->fields_value = $this->getFieldsValueForForm();
-        $helper->languages = $this->context->controller->getLanguages();
-
-        return $helper;
-    }
-
-    /**
-     * @param array $feePlans
-     *
-     * @return array
-     */
-    public function getFieldsValueFromFeePlans($feePlans)
-    {
-        $installmentsPlans = json_decode(SettingsHelper::getFeePlans());
-        $fieldsValue = [];
-        $sortOrder = 1;
-        foreach ($feePlans as $feePlan) {
-            $key = $this->settingsHelper->keyForFeePlan($feePlan);
-
-            $fieldsValue["ALMA_{$key}_ENABLED_ON"] = isset($installmentsPlans->$key->enabled)
-                ? $installmentsPlans->$key->enabled
-                : 0;
-
-            $minAmount = isset($installmentsPlans->$key->min)
-                ? $installmentsPlans->$key->min
-                : $feePlan->min_purchase_amount;
-
-            $fieldsValue["ALMA_{$key}_MIN_AMOUNT"] = (int) round(
-                $this->priceHelper->convertPriceFromCents($minAmount)
-            );
-            $maxAmount = isset($installmentsPlans->$key->max)
-                ? $installmentsPlans->$key->max
-                : $feePlan->max_purchase_amount;
-
-            $fieldsValue["ALMA_{$key}_MAX_AMOUNT"] = (int) $this->priceHelper->convertPriceFromCents($maxAmount);
-
-            $order = isset($installmentsPlans->$key->order)
-                ? $installmentsPlans->$key->order
-                : $sortOrder;
-
-            $fieldsValue["ALMA_{$key}_SORT_ORDER"] = $order;
-
-            ++$sortOrder;
-        }
-
-        return $fieldsValue;
     }
 }
